+++ src/app/routes/routes.tsx
@@ -8,11 +8,16 @@ import AddUserPage from "@/pages/admin/AddUserPage";
 import ProtectedRoute from "../providers/ProtectedRoute";
 import RoleBasedRedirect from "../providers/RoleBasedRedirect";
 import Layout from "../layouts/Layout";
+import PasswordChangePage from "@/pages/PasswordChangePage";
 
 const router = createBrowserRouter([
   {
     path: "/login",
     element: <LoginPage />,
+  },
+  {
+    path: "/passwordchange",
+    element: <PasswordChangePage />,
   },
   {
     path: "/", // 메인 페이지 보호
@@ -27,23 +32,23 @@ const router = createBrowserRouter([
           },
           {
             path: "mypage",
-            element: <RoleBasedRedirect />
+            element: <RoleBasedRedirect />,
           },
           {
             path: "userpage",
-            element: <UserPage />
+            element: <UserPage />,
           },
           {
             path: "adminpage",
-            element: <Outlet />, 
+            element: <Outlet />,
             children: [
               { path: "mytask", element: <MyTaskPage /> },
               { path: "workview", element: <WorkViewPage /> },
-              { path: "adduser", element: <AddUserPage />},
-            ]
-          }
-        ]
-      }, 
+              { path: "adduser", element: <AddUserPage /> },
+            ],
+          },
+        ],
+      },
     ],
   },
 ]);
